// src/types/event.ts
export interface IEvent {
    _id: string;
    title: string;
    description: string;
    date: string;
    location: string;
    owner: string;
    attendees: string;
    likes: string[];
    createdAt: string;
    image: string;
    time: string;
    profilePicture: string;
    name: string;
    followers: string[];
}